+++ addon/utils/information.js
@@ -112,17 +112,17 @@ class Information {
   }
 
   /**
-   Checks if the specified attribute path is a order field.
+   Checks if the specified attribute path is a ordered field.
 
-   @method isDetail
+   @method isOrdered
    @param {String} modelName The name of the model.
    @param {String} attributePath The path to the attribute.
    @returns {Boolean} True if the specified attribute path is a detail field.
    @public
    */
-  isOrder(modelName, attributePath) {
+  isOrdered(modelName, attributePath) {
     let meta = this.getMeta(modelName, attributePath);
-    return meta.isOrder;
+    return meta.isOrdered;
   }
 
   /**
@@ -162,13 +162,13 @@ class Information {
         let attribute = attributes.get(fields[i]);
         if (attribute) {
           let transform = Ember.getOwner(this._store).lookup('transform:' + attribute.type);
-          let order = Ember.isNone(attribute.options) ? false : attribute.options.order;
+          let ordered = Ember.isNone(attribute.options) ? false : attribute.options.ordered;
           return {
             isMaster: false,
             isDetail: false,
             isKey: false,
             isEnum: transform instanceof FlexberryEnum,
-            isOrder: order || false,
+            isOrdered: ordered || false,
             sourceType: transform.get('sourceType'),
             type: attribute.type
           };
@@ -184,7 +184,7 @@ class Information {
             isDetail: isDetail,
             isKey: isMaster,
             isEnum: false,
-            isOrder: false,
+            isOrdered: false,
             type: relationship.type,
             keyType: model.idType || 'guid'
           };
@@ -196,7 +196,7 @@ class Information {
             isDetail: false,
             isKey: true,
             isEnum: false,
-            isOrder: false,
+            isOrdered: false,
             type: 'string',
             keyType: model.idType || 'guid'
           };